--- src/main/java/com/it_academyproject/Domains/Absence.java
@@ -1,66 +1,0 @@
-package com.it_academyproject.Domains;
-
-import org.assertj.core.internal.Dates;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import java.util.Date;
-
-@Entity
-public class Absence {
-	
-	@GeneratedValue(strategy=GenerationType.IDENTITY)	
-	@Id
-	private int id;
-	private Date dateMissing;
-	@ManyToOne (fetch = FetchType.EAGER)
-	@JoinColumn (name="student_id")
-	private MyAppUser userStudent;
-	
-	public Absence() {
-		
-	}
-	public Absence(int id, Date dateMissing) {
-		
-		this.id = id;
-		this.dateMissing = dateMissing;
-	}
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public Date getDateMissing() {
-		return dateMissing;
-	}
-
-	public void setDateMissing(Date dateMissing) {
-		this.dateMissing = dateMissing;
-	}
-
-	public MyAppUser getUserStudent() {
-		return userStudent;
-	}
-
-	public void setUserStudent(MyAppUser userStudent) {
-		this.userStudent = userStudent;
-	}
-
-	@Override
-	public String toString() {
-		return "Absence [id=" + id + ", dateMissing=" + dateMissing + ", user=" + userStudent + "]";
-	}
-
-	
-}
-
-
