--- src/main/java/com/it_academyproject/Domains/Exercice.java
@@ -1,62 +1,0 @@
-package com.it_academyproject.Domains;
-
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-
-@Entity
-public class Exercice {
-	
-	//--------------------------Properties--------------------------------------------------------------
-	
-	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
-	private String name;
-	
-	@ManyToOne(fetch = FetchType.EAGER)
-	@JoinColumn(name="itinerary_id")
-	private Itinerary itinerary;
-	
-	//--------------------------Constructors--------------------------------------------------------------
-
-	public Exercice() {
-	}
-	
-	public Exercice(String name) {
-		this.name = name;
-	}
-
-	//--------------------------Setters/Getters--------------------------------------------------------------------
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public Itinerary getItinerary() {
-		return itinerary;
-	}
-
-	public void setItinerary(Itinerary itinerary) {
-		this.itinerary = itinerary;
-	}
-	
-
-}
