--- src/main/java/com/it_academyproject/Domains/StatusExercice.java
@@ -1,46 +1,0 @@
-package com.it_academyproject.Domains;
-
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class StatusExercice {
-	
-	//--------------------------Properties--------------------------------------------------------------
-	
-	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	private int id;
-	private String name;
-	
-	//--------------------------Constructors--------------------------------------------------------------
-
-	public StatusExercice() {
-	}
-	
-	public StatusExercice(String name) {
-		this.name = name;
-	}
-	
-	//--------------------------Setters/Getters--------------------------------------------------------------------
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-}
