+++ src/main/java/com/it_academyproject/controllers/DataImportController.java
@@ -6,10 +6,12 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.it_academyproject.tools.dataImporter.DataImporter;
 
+import java.util.Map;
+
 
 @RestController
 @RequestMapping("/api/public/")
-public class PublicController
+public class DataImportController
 {
     @Autowired
     DataImporter dataImporter;
@@ -19,12 +21,12 @@ public class PublicController
     public String getMessage()
     {
         JSONObject errorData = new JSONObject();
-        dataImporter.manualCreation();
+        Map<Integer , String> teacherVSitinerary = dataImporter.manualCreation();
         dataImporter.importAlumnesActius();
-        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 0 , 1) );
-        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 1 , 2) );
-        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 2 , 3) );
-        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 3 , 4) );
+        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 0 , 1 , teacherVSitinerary.get(1) ) );
+        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 1 , 2 , teacherVSitinerary.get(2) ) );
+        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 2 , 3 , teacherVSitinerary.get(3) ) );
+        errorData.put( "Exercises" , dataImporter.importEjerciciosporalumno( 3 , 4 , teacherVSitinerary.get(4) ) );
         dataImporter.importTaules();
 
         return errorData.toString();
