+++ src/main/java/com/it_academyproject/exceptions/GenericResponse.java
@@ -1,4 +1,4 @@
-package com.it_academyproject.Exceptions;
+package com.it_academyproject.exceptions;
 
 public class GenericResponse {
     private String message;
