+++ src/main/java/com/it_academyproject/exceptions/WrongEmailPassword.java
@@ -1,4 +1,4 @@
-package com.it_academyproject.Exceptions;
+package com.it_academyproject.exceptions;
 
 public class WrongEmailPassword extends Exception
 {
