+++ src/main/java/com/it_academyproject/repositories/ItineraryRepository.java
@@ -1,7 +1,6 @@
 package com.it_academyproject.repositories;
 
-import com.it_academyproject.Domains.Itinerary;
-import com.it_academyproject.Domains.Role;
+import com.it_academyproject.domains.Itinerary;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
