+++ src/main/java/com/it_academyproject/repositories/StatusExerciceRepository.java
@@ -1,7 +1,6 @@
 package com.it_academyproject.repositories;
 
-import com.it_academyproject.Domains.Exercice;
-import com.it_academyproject.Domains.StatusExercice;
+import com.it_academyproject.domains.StatusExercice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface StatusExerciceRepository extends JpaRepository<StatusExercice, Integer>
