+++ src/main/java/com/it_academyproject/repositories/UserExerciceRepository.java
@@ -1,8 +1,8 @@
 package com.it_academyproject.repositories;
 
-import com.it_academyproject.Domains.Exercice;
-import com.it_academyproject.Domains.MyAppUser;
-import com.it_academyproject.Domains.UserExercice;
+import com.it_academyproject.domains.Exercice;
+import com.it_academyproject.domains.MyAppUser;
+import com.it_academyproject.domains.UserExercice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserExerciceRepository extends JpaRepository <UserExercice , Integer>
